+++ src/components/app/media/index.ts
@@ -1,0 +1,1 @@
+export { default as FileGallery } from './FileGallery';