+++ src/components/app/typography/Paragraph/Paragraph-types.ts
@@ -1,0 +1,11 @@
+type Color = 'black' | 'white';
+
+export interface ViewProps {
+  /** - */
+  children: React.ReactNode | string;
+  color?: Color;
+}
+
+export interface StylesProps {
+  color?: Color;
+}