+++ src/components/app/typography/index.ts
@@ -1,0 +1,5 @@
+export { default as Detail } from './Detail';
+export { default as Heading } from './Heading';
+export { default as Paragraph } from './Paragraph';
+export { default as Subheading } from './Subheading';
+export { default as Title } from './Title';